requirements.txt
# Core dependencies for Etsy flow replication
curl-cffi>=0.7.0
playwright>=1.40.0
playwright-stealth>=1.0.6
requests>=2.31.0
cloudscraper>=1.2.71
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dotenv>=1.0.0
colorama>=0.4.6